fix: gpg-agent as ssh-agent using yubikey
from the gpg docs:
```
SSH has no way to tell the gpg-agent what terminal or X display it is
running on. So when remotely logging into a box where a gpg-agent with
SSH support is running, the pinentry will get popped up on whatever
display the gpg-agent has been started. To solve this problem you may
issue the command
```

https://www.gnupg.org/documentation/manuals/gnupg/Common-Problems.html